MiliSec: change to u64

+++ src/lib.rs
@@ -29,4 +29,4 @@ mod state;
 pub mod test_utils;
 mod timer;
 
-pub type MiliSec = u32;
+pub type MiliSec = u64;
